deletion of duplicated codes for redo and undo + 1 modification for AbstractCopyCutAction
