+++ jhotdraw-actions/src/main/java/org/jhotdraw/action/edit/RedoAction.java
@@ -32,85 +32,13 @@ import org.jhotdraw.util.*;
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class RedoAction extends AbstractViewAction {
-
-    private static final long serialVersionUID = 1L;
+public class RedoAction extends AbstractRedoUndoAction {
     public static final String ID = "edit.redo";
-    private ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.action.Labels");
-    private PropertyChangeListener redoActionPropertyListener = new PropertyChangeListener() {
-        @Override
-        public void propertyChange(PropertyChangeEvent evt) {
-            String name = evt.getPropertyName();
-            if ((name == null && AbstractAction.NAME == null) || (name != null && name.equals(AbstractAction.NAME))) {
-                putValue(AbstractAction.NAME, evt.getNewValue());
-            } else if ("enabled".equals(name)) {
-                updateEnabledState();
-            }
-        }
-    };
 
     /**
      * Creates a new instance.
      */
     public RedoAction(Application app, View view) {
-        super(app, view);
-        labels.configureAction(this, ID);
-    }
-
-    protected void updateEnabledState() {
-        boolean isEnabled = false;
-        Action realRedoAction = getRealRedoAction();
-        if (realRedoAction != null && realRedoAction != this) {
-            isEnabled = realRedoAction.isEnabled();
-        }
-        setEnabled(isEnabled);
-    }
-
-    @Override
-    protected void updateView(View oldValue, View newValue) {
-        super.updateView(oldValue, newValue);
-        if (newValue != null
-                && newValue.getActionMap().get(ID) != null
-                && newValue.getActionMap().get(ID) != this) {
-            putValue(AbstractAction.NAME, newValue.getActionMap().get(ID).
-                    getValue(AbstractAction.NAME));
-            updateEnabledState();
-        }
-    }
-
-    /**
-     * Installs listeners on the view object.
-     */
-    @Override
-    protected void installViewListeners(View p) {
-        super.installViewListeners(p);
-        Action redoActionInView = p.getActionMap().get(ID);
-        if (redoActionInView != null && redoActionInView != this) {
-            redoActionInView.addPropertyChangeListener(redoActionPropertyListener);
-        }
-    }
-
-    /**
-     * Installs listeners on the view object.
-     */
-    @Override
-    protected void uninstallViewListeners(View p) {
-        super.uninstallViewListeners(p);
-        Action redoActionInView = p.getActionMap().get(ID);
-        if (redoActionInView != null && redoActionInView != this) {
-            redoActionInView.removePropertyChangeListener(redoActionPropertyListener);
-        }
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        Action realAction = getRealRedoAction();
-        if (realAction != null && realAction != this) {
-            realAction.actionPerformed(e);
-        }
-    }
-
-    private Action getRealRedoAction() {
-        return (getActiveView() == null) ? null : getActiveView().getActionMap().get(ID);
+        super(app, view, ID);
     }
 }
