+++ jhotdraw-actions/src/main/java/org/jhotdraw/action/edit/UndoAction.java
@@ -31,85 +31,13 @@ import org.jhotdraw.util.*;
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class UndoAction extends AbstractViewAction {
-
-    private static final long serialVersionUID = 1L;
+public class UndoAction extends AbstractRedoUndoAction {
     public static final String ID = "edit.undo";
-    private ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.action.Labels");
-    private PropertyChangeListener redoActionPropertyListener = new PropertyChangeListener() {
-        @Override
-        public void propertyChange(PropertyChangeEvent evt) {
-            String name = evt.getPropertyName();
-            if ((name == null && AbstractAction.NAME == null) || (name != null && name.equals(AbstractAction.NAME))) {
-                putValue(AbstractAction.NAME, evt.getNewValue());
-            } else if ("enabled".equals(name)) {
-                updateEnabledState();
-            }
-        }
-    };
 
     /**
      * Creates a new instance.
      */
     public UndoAction(Application app, View view) {
-        super(app, view);
-        labels.configureAction(this, ID);
-    }
-
-    protected void updateEnabledState() {
-        boolean isEnabled = false;
-        Action realAction = getRealUndoAction();
-        if (realAction != null && realAction != this) {
-            isEnabled = realAction.isEnabled();
-        }
-        setEnabled(isEnabled);
-    }
-
-    @Override
-    protected void updateView(View oldValue, View newValue) {
-        super.updateView(oldValue, newValue);
-        if (newValue != null
-                && newValue.getActionMap().get(ID) != null
-                && newValue.getActionMap().get(ID) != this) {
-            putValue(AbstractAction.NAME, newValue.getActionMap().get(ID).
-                    getValue(AbstractAction.NAME));
-            updateEnabledState();
-        }
-    }
-
-    /**
-     * Installs listeners on the view object.
-     */
-    @Override
-    protected void installViewListeners(View p) {
-        super.installViewListeners(p);
-        Action undoActionInView = p.getActionMap().get(ID);
-        if (undoActionInView != null && undoActionInView != this) {
-            undoActionInView.addPropertyChangeListener(redoActionPropertyListener);
-        }
-    }
-
-    /**
-     * Installs listeners on the view object.
-     */
-    @Override
-    protected void uninstallViewListeners(View p) {
-        super.uninstallViewListeners(p);
-        Action undoActionInView = p.getActionMap().get(ID);
-        if (undoActionInView != null && undoActionInView != this) {
-            undoActionInView.removePropertyChangeListener(redoActionPropertyListener);
-        }
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        Action realUndoAction = getRealUndoAction();
-        if (realUndoAction != null && realUndoAction != this) {
-            realUndoAction.actionPerformed(e);
-        }
-    }
-
-    private Action getRealUndoAction() {
-        return (getActiveView() == null) ? null : getActiveView().getActionMap().get(ID);
+        super(app, view, ID);
     }
 }
